app/schemas/chat.py
from pydantic import BaseModel
from typing import Optional

class ChatRequest(BaseModel):
    message: str 
    user_id: Optional[str] = None

class ChatResponse(BaseModel):
    reply: str
    user_id: Optional[str] = None