requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy>=2.0
asyncpg
redis
python-dotenv
twilio
pytest
pytest-asyncio
